comple side navbar design and also add some extra routes
