+++ app/dashboard/settings/page.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const page = () => {
+  return <div>this is setting page</div>;
+};
+
+export default page;
